Frontend/Demo3_Frontend/Demo4/app/src/androidTest/java/com/example/androidexample/LevelEditorTest.java
package com.example.androidexample;

import android.view.MotionEvent;
import android.view.View;

import androidx.test.core.app.ActivityScenario;
import androidx.test.espresso.Espresso;
import androidx.test.espresso.action.CoordinatesProvider;
import androidx.test.espresso.action.GeneralClickAction;
import androidx.test.espresso.action.GeneralLocation;
import androidx.test.espresso.action.Press;
import androidx.test.espresso.action.Tap;
import androidx.test.espresso.action.ViewActions;
import androidx.test.espresso.assertion.ViewAssertions;
import androidx.test.espresso.matcher.ViewMatchers;
import androidx.test.ext.junit.rules.ActivityScenarioRule;
import androidx.test.ext.junit.runners.AndroidJUnit4;

import org.junit.Rule;
import org.junit.Test;
import org.junit.runner.RunWith;

@RunWith(AndroidJUnit4.class)
public class LevelEditorTest {

    @Rule
    public ActivityScenarioRule<LevelEditor> activityRule = new ActivityScenarioRule<>(LevelEditor.class);


    @Test
    public void testRemoveButtonClick() {
        // Click on the remove button
        Espresso.onView(ViewMatchers.withId(R.id.remove)).perform(ViewActions.click());

        // Add assertions to verify the expected behavior of the remove button
        // For example, you can check if a specific action has been performed when the button is clicked
        // (e.g., if a specific method is called in the activity)
    }
    @Test
    public void testUndo() {
        // Trigger the undo method
        // You may need to adjust the expected values based on your implementation
        // Ensure that your UI elements are in a state where you can observe the changes
        Espresso.onView(ViewMatchers.withId(R.id.undo)).perform(ViewActions.click());

        // Check if the expected response is displayed or if UI state has changed
    }
    @Test
    public void testBlueButtonClick() {
        // Click on the remove button
        Espresso.onView(ViewMatchers.withId(R.id.blue)).perform(ViewActions.click());

        Espresso.onView(ViewMatchers.withId(R.id.surfaceView)).perform(ViewActions.click());

        // Add assertions to verify the expected behavior of the remove button
        // For example, you can check if a specific action has been performed when the button is clicked
        // (e.g., if a specific method is called in the activity)
    }
    @Test
    public void testYellowButtonClick() {
        // Click on the remove button
        Espresso.onView(ViewMatchers.withId(R.id.yellow)).perform(ViewActions.click());

        Espresso.onView(ViewMatchers.withId(R.id.surfaceView)).perform(ViewActions.click());

        // Add assertions to verify the expected behavior of the remove button
        // For example, you can check if a specific action has been performed when the button is clicked
        // (e.g., if a specific method is called in the activity)
    }
    @Test
    public void testGreenButtonClick() {
        // Click on the remove button
        Espresso.onView(ViewMatchers.withId(R.id.green)).perform(ViewActions.click());

        Espresso.onView(ViewMatchers.withId(R.id.surfaceView)).perform(ViewActions.click());

        // Add assertions to verify the expected behavior of the remove button
        // For example, you can check if a specific action has been performed when the button is clicked
        // (e.g., if a specific method is called in the activity)
    }
    @Test
    public void testredButtonClick() {
        // Click on the remove button
        Espresso.onView(ViewMatchers.withId(R.id.red)).perform(ViewActions.click());

        Espresso.onView(ViewMatchers.withId(R.id.surfaceView)).perform(ViewActions.click());

        // Add assertions to verify the expected behavior of the remove button
        // For example, you can check if a specific action has been performed when the button is clicked
        // (e.g., if a specific method is called in the activity)
    }
    @Test
    public void testmagentaButtonClick() {
        // Click on the remove button
        Espresso.onView(ViewMatchers.withId(R.id.magenta)).perform(ViewActions.click());

        Espresso.onView(ViewMatchers.withId(R.id.surfaceView)).perform(ViewActions.click());

        // Add assertions to verify the expected behavior of the remove button
        // For example, you can check if a specific action has been performed when the button is clicked
        // (e.g., if a specific method is called in the activity)
    }
    @Test
    public void testremoveBthClick() {
        // Click on the remove button
        Espresso.onView(ViewMatchers.withId(R.id.remove)).perform(ViewActions.click());

        Espresso.onView(ViewMatchers.withId(R.id.surfaceView)).perform(ViewActions.click());

        // Add assertions to verify the expected behavior of the remove button
        // For example, you can check if a specific action has been performed when the button is clicked
        // (e.g., if a specific method is called in the activity)
    }
    @Test
    public void testPhysicsInteraction() {
        // Click on the remove button
        Espresso.onView(ViewMatchers.withId(R.id.green)).perform(ViewActions.click());

        // Perform a click at a specific spot on the SurfaceView
        float x1 = 540;
        float y1 = 1006;
        Espresso.onView(ViewMatchers.withId(R.id.surfaceView)).perform(ViewActions.actionWithAssertions(
                new GeneralClickAction(Tap.SINGLE, new CoordinatesProvider() {
                    @Override
                    public float[] calculateCoordinates(View view) {
                        int[] locationOnScreen = new int[2];
                        view.getLocationOnScreen(locationOnScreen);
                        float x = locationOnScreen[0] + x1;
                        float y = locationOnScreen[1] + y1;
                        return new float[]{x, y};
                    }
                }, Press.THUMB)));

        // Wait for the button to be clicked
        try {
            Thread.sleep(1000);
        } catch (InterruptedException e) {
            e.printStackTrace();
        }

        // Click on the yellow button
        Espresso.onView(ViewMatchers.withId(R.id.yellow)).perform(ViewActions.click());

        // Perform a click right above the first click
        float x2 = 540; // Same x-coordinate
        float y2 = 960; /* specify the vertical distance between the clicks */
        Espresso.onView(ViewMatchers.withId(R.id.surfaceView)).perform(ViewActions.actionWithAssertions(
                new GeneralClickAction(Tap.SINGLE, new CoordinatesProvider() {
                    @Override
                    public float[] calculateCoordinates(View view) {
                        int[] locationOnScreen = new int[2];
                        view.getLocationOnScreen(locationOnScreen);
                        float x = locationOnScreen[0] + x2;
                        float y = locationOnScreen[1] + y2;
                        return new float[]{x, y};
                    }
                }, Press.THUMB)));

        try {
            Thread.sleep(2000);
        } catch (InterruptedException e) {
            e.printStackTrace();
        }
        // Add assertions to verify the expected behavior of the buttons
        // For example, you can check if a specific action has been performed when the buttons are clicked
        // (e.g., if a specific method is called in the activity)
    }
    @Test
    public void testAntiGravityInteraction() {
        // Click on the remove button
        Espresso.onView(ViewMatchers.withId(R.id.magenta)).perform(ViewActions.click());

        // Perform a click at a specific spot on the SurfaceView
        float x1 = 540;
        float y1 = 1006;
        Espresso.onView(ViewMatchers.withId(R.id.surfaceView)).perform(ViewActions.actionWithAssertions(
                new GeneralClickAction(Tap.SINGLE, new CoordinatesProvider() {
                    @Override
                    public float[] calculateCoordinates(View view) {
                        int[] locationOnScreen = new int[2];
                        view.getLocationOnScreen(locationOnScreen);
                        float x = locationOnScreen[0] + x1;
                        float y = locationOnScreen[1] + y1;
                        return new float[]{x, y};
                    }
                }, Press.THUMB)));

        // Wait for the button to be clicked
        try {
            Thread.sleep(1000);
        } catch (InterruptedException e) {
            e.printStackTrace();
        }

        // Click on the yellow button
        Espresso.onView(ViewMatchers.withId(R.id.yellow)).perform(ViewActions.click());

        // Perform a click right above the first click
        float x2 = 540; // Same x-coordinate
        float y2 = 960; /* specify the vertical distance between the clicks */
        Espresso.onView(ViewMatchers.withId(R.id.surfaceView)).perform(ViewActions.actionWithAssertions(
                new GeneralClickAction(Tap.SINGLE, new CoordinatesProvider() {
                    @Override
                    public float[] calculateCoordinates(View view) {
                        int[] locationOnScreen = new int[2];
                        view.getLocationOnScreen(locationOnScreen);
                        float x = locationOnScreen[0] + x2;
                        float y = locationOnScreen[1] + y2;
                        return new float[]{x, y};
                    }
                }, Press.THUMB)));

        try {
            Thread.sleep(2000);
        } catch (InterruptedException e) {
            e.printStackTrace();
        }
        // Add assertions to verify the expected behavior of the buttons
        // For example, you can check if a specific action has been performed when the buttons are clicked
        // (e.g., if a specific method is called in the activity)
    }

    @Test
    public void testSurfaceViewInteraction() {
        // Simulate a touch event on the SurfaceView
        Espresso.onView(ViewMatchers.withId(R.id.surfaceView)).perform(ViewActions.click());

        // Add assertions to verify the expected behavior of the SurfaceView
        // For example, check if the physicsSimulator has been updated or if the level data has been sent to the server
    }

    @Test
    public void testPostLevelButtonClick() {
        // Click on the "Post Level" button
        Espresso.onView(ViewMatchers.withId(R.id.post)).perform(ViewActions.click());


        // Add assertions to verify the expected behavior when the button is clicked
        // For example, check if the saveSimulationData method has been called
    }



    // Add more tests based on your UI interactions and expected behavior
}
